use theme combobox instead of dark mode toggle add system theme to automatically select system dark mode (default) fix some dialogs looking weird in dark mode fix tables contain ghost items in unused rows (untested in Windows/macOS) implement better caught detection item scoring, based of item origin spawn position in comparison to last bobber position add client brand fishingbot

+++ src/main/java/systems/kinau/fishingbot/event/play/SpawnEntityEvent.java
@@ -14,9 +14,9 @@ public class SpawnEntityEvent extends Event {
 
     @Getter private int id;
     @Getter private int type;
-    @Getter private int x;
-    @Getter private int y;
-    @Getter private int z;
+    @Getter private double x;
+    @Getter private double y;
+    @Getter private double z;
     @Getter private byte yaw;
     @Getter private byte pitch;
     @Getter private int objectData;
